requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
pdfplumber==0.11.4
PyMuPDF==1.24.10
ollama==0.3.3
pydantic==2.8.0
python-dotenv==1.0.1
httpx==0.27.0
Pillow==10.4.0
pytesseract==0.3.13